refactor: OrganizationResponse 파일 경로 변경 및 필드명 수정
