+++ android/app/src/main/java/com/daedan/festabook/data/datasource/remote/organization/FestivalDataSourceImpl.kt
@@ -1,7 +1,7 @@
 package com.daedan.festabook.data.datasource.remote.organization
 
 import com.daedan.festabook.data.datasource.remote.ApiResult
-import com.daedan.festabook.data.model.response.OrganizationResponse
+import com.daedan.festabook.data.model.response.festival.OrganizationResponse
 import com.daedan.festabook.data.service.FestivalService
 
 class FestivalDataSourceImpl(
