+++ android/app/src/main/java/com/daedan/festabook/data/service/FestivalService.kt
@@ -1,6 +1,6 @@
 package com.daedan.festabook.data.service
 
-import com.daedan.festabook.data.model.response.OrganizationResponse
+import com.daedan.festabook.data.model.response.festival.OrganizationResponse
 import com.daedan.festabook.data.model.response.organization.OrganizationGeographyResponse
 import retrofit2.Response
 import retrofit2.http.GET
